CAMEL-16197: camel-cxf - Should ensure CXF attachment that are cached to temp disk should be closed/deleted when UoW is done so CXF does not leak files.
