+++ components/camel-cxf/src/main/java/org/apache/camel/component/cxf/DefaultCxfBinding.java
@@ -23,6 +23,7 @@ import java.lang.reflect.Modifier;
 import java.nio.charset.Charset;
 import java.security.Principal;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
@@ -31,6 +32,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.TreeMap;
 
+import javax.activation.DataHandler;
 import javax.security.auth.Subject;
 import javax.xml.XMLConstants;
 import javax.xml.namespace.QName;
@@ -46,6 +48,7 @@ import org.w3c.dom.Node;
 import org.apache.camel.Exchange;
 import org.apache.camel.ExchangePattern;
 import org.apache.camel.ExchangePropertyKey;
+import org.apache.camel.ExtendedExchange;
 import org.apache.camel.attachment.AttachmentMessage;
 import org.apache.camel.attachment.DefaultAttachment;
 import org.apache.camel.component.cxf.common.header.CxfHeaderHelper;
@@ -54,6 +57,8 @@ import org.apache.camel.component.cxf.util.ReaderInputStream;
 import org.apache.camel.spi.HeaderFilterStrategy;
 import org.apache.camel.spi.HeaderFilterStrategyAware;
 import org.apache.camel.support.ExchangeHelper;
+import org.apache.camel.support.SynchronizationAdapter;
+import org.apache.camel.util.IOHelper;
 import org.apache.camel.util.ObjectHelper;
 import org.apache.cxf.attachment.AttachmentImpl;
 import org.apache.cxf.binding.soap.Soap11;
@@ -201,6 +206,36 @@ public class DefaultCxfBinding implements CxfBinding, HeaderFilterStrategyAware 
                         createCamelAttachment(attachment));
             }
         }
+        addAttachmentFileCloseUoW(camelExchange, cxfExchange);
+    }
+
+    /**
+     * CXF may cache attachments in the filesystem temp folder. The files may leak if they were not used. Add a cleanup
+     * handler to remove the attachments after message processing.
+     *
+     * @param camelExchange
+     * @param cxfExchange
+     */
+    private void addAttachmentFileCloseUoW(Exchange camelExchange, org.apache.cxf.message.Exchange cxfExchange) {
+        camelExchange.adapt(ExtendedExchange.class).addOnCompletion(new SynchronizationAdapter() {
+            @Override
+            public void onDone(org.apache.camel.Exchange exchange) {
+                Collection<Attachment> atts = cxfExchange.getInMessage().getAttachments();
+                if (atts != null) {
+                    for (Attachment att : atts) {
+                        DataHandler dh = att.getDataHandler();
+                        if (dh != null) {
+                            try {
+                                InputStream is = dh.getInputStream();
+                                IOHelper.close(is);
+                            } catch (Exception e) {
+                                // ignore
+                            }
+                        }
+                    }
+                }
+            }
+        });
     }
 
     private DefaultAttachment createCamelAttachment(Attachment attachment) {
@@ -319,6 +354,7 @@ public class DefaultCxfBinding implements CxfBinding, HeaderFilterStrategyAware 
                         createCamelAttachment(attachment));
             }
         }
+        addAttachmentFileCloseUoW(camelExchange, cxfExchange);
     }
 
     /**
